backend/services/payments/requirements.txt
firebase-admin
flask
gunicorn
flask-cors
google-cloud-secret-manager
stripe
